requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
google-auth
google-auth-oauthlib
google-auth-httplib2
itsdangerous